perf: Set the DSU PMU to be readable from any CPU
Usually, the perf core assumes that the PMU is attached to
a particular processor core and ignores the event to be read
if that particular CPU is idle. However, the DSU PMU isn't
attached to any core. In fact, it is common for all the cores
and the event can be safely read from any cpu. As a result,
initialize the event to achieve the same.

Change-Id: I5037e749019471476369bdf818602faa70de26a6
Signed-off-by: Raghavendra Rao Ananta <rananta@codeaurora.org>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>