arm64: dts: qcom: Make QUP SE devices dma coherent
This change adds memory attributes “dma-coherent” defining the
io-coherent access by QUP SE Devices which will be checked by
SMMU for address translation and memory attributes. This removes
the CMOs and helps in faster translation. This is as part of the
FR62360.

IOMMU page tables should already be IO-coherent from Lahaina
onwards, so we don’t need to define qcom,iommu-pagetable as
coherent.

Change-Id: I4d76e6e36871a3665c5367c1c34c80ae3a9f6db2
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>