techpack: disp: Skip heavy autorefresh checks when it's not enabled
These heavy checks for seeing if autorefresh is enabled are unneeded
when the autorefresh config is disabled. These checks are performed on
every display commit and show up as using a significant amount of CPU
time in perf top. Skip them when it's unnecessary in order to improve
display rendering performance.

Change-Id: Ib0078894dda5f7267673d1047404ab2c85107379
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>